WordsRestService/requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
colorama==0.4.0
coverage==4.5.1
Django==1.11.16
djangorestframework==3.9.0
Faker==0.9.1
mixer==6.1.3
more-itertools==4.3.0
pip==18.1
pluggy==0.8.0
py==1.7.0
pytest==3.9.1
pytest-cov==2.6.0
pytest-django==3.4.3
python-dateutil==2.7.3
pytz==2018.5
setuptools==39.0.1
six==1.11.0
text-unidecode==1.2
